worker: add generic error to the OSBuildJobResult
Some job failures cannot be described in a structured way. This commit
introduces a generic error to the OSBuildJobResult struct for these kinds
of errors. Also, it immediately abuses the generic error to report back
the target errors. Hold on, we will get there and make things pretty.

RunJob no longer returns an error. This is not very idiomatic but I believe
the approach with returning a complex result struct is much more appropriate
in this use-case.