spec: obsolete lorax-composer
osbuild-composer is a drop-in replacement of lorax-composer, and
side-by-side installation is not supported. It is still possible to
explicitly install lorax-composer and exclude osbuild-composer, but the
default is to replace lorax with osbuild on upgrades.

This change is for now only for RHEL8.3, but we should consider doing
the same for F33.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1836844

Signed-off-by: Tom Gundersen <teg@jklm.no>