schutzbot/mockbuild: test against osbuild in target release by default
To make sure we test against what we will release with, osbuild from the
target distro.

The only exception is in RHEL8.3, where we are not rebasing any longer,
but expect users who use new osbulid-composer from upstream to use it
with new osbuild from upstream too. In this case, build osbuild from
the submodule.

In the future, we should rely on the osbuild package being built by
osbuild CI and pull it directly from the repo provided there.

Signed-off-by: Tom Gundersen <teg@jklm.no>