introduce cmd/osbuild-base-test-runner
This executable will be used for running base tests agains remote
osbuild-composer. The idea is to spawn VM in AWS using this executable,
forward the Weldr API socket to the local machine, and finally run any
test as if the osbuild-composer was running locally.

The great news about this approach is that we already have our own
awesome implementation of Go "context managers" for booting a VM in AWS.
This patch simply copies & pastes the already existing implementation,
because it is not available in the common (internal) package.