base-tests: copy two already existing scripts for future modifications
The base test runner needs a preceding stage that uploads an image with
osbuild-composer enabled into EC2. This will be the job of
"upload-to-aws.sh".

It also needs a script to run the Go executable itself, this one will be
similar to the script that runs base tests as of today.

Both files are unmodified copies and both will be tweaked in following
commits.