schutzbot: create "sources" configuration from the S3 repo
In the following commits I'll introduce image creation using
osbuild-composer where I'll use this repository. The problem is that
composer-cli does not accept the repo file as a source format, it
requires TOML specification. This patch stores the same configuration,
but in TOML.

(stashing in Jenkinsfile will be introduced later)